# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * UFC Card Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface UfcEvent {
  id: string;
  name: string;
  date: string;
  /** @nullable */
  venue?: string | null;
  /** @nullable */
  location?: string | null;
}

export interface FighterSummary {
  name: string;
  record: string;
  /** @nullable */
  ufcStatsId?: string | null;
}

export interface FightCard {
  id: string;
  weightClass: string;
  order?: number;
  isMain?: boolean;
  fighterA: FighterSummary;
  fighterB: FighterSummary;
}

export interface EventCard {
  id: string;
  name: string;
  date: string;
  /** @nullable */
  venue?: string | null;
  /** @nullable */
  location?: string | null;
  fights: FightCard[];
}

export interface FighterStats {
  name: string;
  record: string;
  /** @nullable */
  reach?: string | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  stance?: string | null;
  style?: string;
  /** @nullable */
  slpm?: number | null;
  /** @nullable */
  strAcc?: number | null;
  /** @nullable */
  strDef?: number | null;
  /** @nullable */
  tdAvg?: number | null;
  /** @nullable */
  tdAcc?: number | null;
  /** @nullable */
  tdDef?: number | null;
  /** @nullable */
  subAvg?: number | null;
  recentForm?: string[];
  strengths?: string[];
  weaknesses?: string[];
}

export interface CommonOpponent {
  opponent: string;
  resultA: string;
  resultB: string;
  /** @nullable */
  methodA?: string | null;
  /** @nullable */
  methodB?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface FightOdds {
  fighterA: string;
  fighterB: string;
  book?: string;
}

export type LeanConfidence = typeof LeanConfidence[keyof typeof LeanConfidence];


export const LeanConfidence = {
  strong: 'strong',
  lean: 'lean',
  'toss-up': 'toss-up',
} as const;

export interface Lean {
  fighter: string;
  confidence: LeanConfidence;
  reasoning: string;
  keyEdges?: string[];
  riskFactors?: string[];
}

export interface Source {
  label: string;
  url: string;
}

export interface FightAnalysis {
  fightId: string;
  weightClass?: string;
  fighterA: FighterStats;
  fighterB: FighterStats;
  commonOpponents?: CommonOpponent[];
  odds?: FightOdds | null;
  lean: Lean;
  /** @nullable */
  styleMatchup?: string | null;
  sources?: Source[];
}