# lib/api-zod/src/generated/types/commonOpponent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * UFC Card Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface CommonOpponent {
  opponent: string;
  resultA: string;
  resultB: string;
  /** @nullable */
  methodA?: string | null;
  /** @nullable */
  methodB?: string | null;
  /** @nullable */
  notes?: string | null;
}