# lib/api-zod/src/generated/types/fightAnalysis.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * UFC Card Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { CommonOpponent } from './commonOpponent';
import type { FighterStats } from './fighterStats';
import type { FightOdds } from './fightOdds';
import type { Lean } from './lean';
import type { Source } from './source';

export interface FightAnalysis {
  fightId: string;
  weightClass?: string;
  fighterA: FighterStats;
  fighterB: FighterStats;
  commonOpponents?: CommonOpponent[];
  odds?: FightOdds | null;
  lean: Lean;
  /** @nullable */
  styleMatchup?: string | null;
  sources?: Source[];
}